moved numba jits to molecule.util. minor updates to other stuff
